[SUS-47] 임시 저장
